moved GenericViewModel
