+++ app/src/main/java/org/commonvoice/saverio_lib/viewmodels/GenericViewModel.kt
@@ -1,9 +1,9 @@
-package org.commonvoice.saverio
+package org.commonvoice.saverio_lib.viewmodels
 
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class GenericViewModel:ViewModel() {
+class GenericViewModel: ViewModel() {
     var fromFragment = MutableLiveData<String>()
     fun updateFromFragment(value: String) {
         fromFragment.value = value
